:recycle: (refactor) move register user handler to onboarding
