+++ apps/conv-lm-backend/src/app/functions/bot/registerUser.function.ts
@@ -1,4 +1,4 @@
-import { RegisterEndUserHandler } from "@app/functions/conversations/progression";
+import { RegisterEndUserHandler } from "@app/functions/conversations/onboarding";
 import { RestRegistrar } from "@ngfi/functions";
 
 import { GCPFunction } from "../../../function";
