--- libs/functions/conversations/progression/src/index.ts
@@ -1,1 +1,0 @@
-export * from './lib/register-user.handler';
